requirements.txt
Flask
Flask-Session
psycopg2-binary
SQLAlchemy
requests
datetime
werkzeug.security
functools
csv